Fixed warning

+++ DirectXTex/DirectXTexTGA.cpp
@@ -1934,7 +1934,7 @@ HRESULT DirectX::LoadFromTGAFile(
         }
 
 #ifdef WIN32
-        if (!ReadFile(hFile.get(), temp.get(), remaining, &bytesRead, nullptr))
+        if (!ReadFile(hFile.get(), temp.get(), static_cast<DWORD>(remaining), &bytesRead, nullptr))
         {
             image.Release();
             return HRESULT_FROM_WIN32(GetLastError());
